KTOR-7075: changes according to review
