+++ ktor-utils/build.gradle.kts
@@ -27,7 +27,7 @@ kotlin {
         commonMain.dependencies {
             api(projects.ktorIo)
             api(libs.kotlinx.serialization.core)
-            api(libs.zstd.jni)
+            implementation(libs.zstd.jni)
         }
         commonTest.dependencies {
             api(projects.ktorTestDispatcher)
