+++ ktor-utils/jvm/src/io/ktor/util/EncodersJvm.kt
@@ -4,9 +4,12 @@
 
 package io.ktor.util
 
+import com.github.luben.zstd.ZstdCompressCtx
+import com.github.luben.zstd.ZstdDecompressCtx
 import io.ktor.util.cio.*
 import io.ktor.utils.io.*
 import io.ktor.utils.io.core.*
+import io.ktor.utils.io.pool.ObjectPool
 import kotlinx.coroutines.*
 import java.nio.*
 import java.util.zip.*
@@ -72,17 +75,91 @@ public val GZip: Encoder = object : Encoder {
  *
  * [Report a problem](https://ktor.io/feedback/?fqname=io.ktor.util.Zstd)
  */
-public class Zstd(private val compressionLevel: Int = 3) : Encoder {
+public class Zstd(private val compressionLevel: Int) : Encoder {
+    @OptIn(DelicateCoroutinesApi::class)
     override fun encode(source: ByteReadChannel, coroutineContext: CoroutineContext): ByteReadChannel =
-        source.encoded(coroutineContext = coroutineContext, compressionLevel = compressionLevel)
+        GlobalScope.writer(coroutineContext, autoFlush = true) {
+            source.encodeTo(channel, KtorDefaultPool, compressionLevel)
+        }.channel
 
+    @OptIn(DelicateCoroutinesApi::class)
     override fun encode(source: ByteWriteChannel, coroutineContext: CoroutineContext): ByteWriteChannel =
-        source.encoded(coroutineContext = coroutineContext, compressionLevel = compressionLevel)
+        GlobalScope.reader(coroutineContext, autoFlush = true) {
+            channel.encodeTo(source, KtorDefaultPool, compressionLevel)
+        }.channel
 
+    @OptIn(DelicateCoroutinesApi::class)
     override fun decode(
         source: ByteReadChannel,
         coroutineContext: CoroutineContext
-    ): ByteReadChannel = source.decoded(coroutineContext = coroutineContext)
+    ): ByteReadChannel = GlobalScope.writer(coroutineContext) {
+        source.decodeTo(channel, KtorDefaultPool)
+    }.channel
+
+    private suspend fun ByteReadChannel.decodeTo(
+        destination: ByteWriteChannel,
+        pool: ObjectPool<ByteBuffer> = KtorDefaultPool
+    ) {
+        val inputBuf = pool.borrow()
+        val outputBuf = pool.borrow()
+        val ctx = ZstdDecompressCtx()
+
+        try {
+            while (!isClosedForRead) {
+                inputBuf.clear()
+                val bytesRead = readAvailable(inputBuf)
+                if (bytesRead <= 0) continue
+
+                val decompressedSize = ctx.decompressByteArray(
+                    outputBuf.array(),
+                    0,
+                    outputBuf.capacity(),
+                    inputBuf.array(),
+                    0,
+                    bytesRead
+                )
+
+                destination.writeFully(outputBuf.array(), 0, decompressedSize)
+            }
+        } finally {
+            ctx.close()
+            pool.recycle(inputBuf)
+            pool.recycle(outputBuf)
+        }
+    }
+
+    private suspend fun ByteReadChannel.encodeTo(
+        destination: ByteWriteChannel,
+        pool: ObjectPool<ByteBuffer> = KtorDefaultPool,
+        compressionLevel: Int,
+    ) {
+        val inputBuf = pool.borrow()
+        val outputBuf = pool.borrow()
+        val ctx = ZstdCompressCtx().apply { setLevel(compressionLevel) }
+
+        try {
+            while (!isClosedForRead) {
+                inputBuf.clear()
+                val bytesRead = readAvailable(inputBuf)
+                if (bytesRead <= 0) continue
+
+                val compressedSize = ctx.compressByteArray(
+                    outputBuf.array(),
+                    0,
+                    outputBuf.capacity(),
+                    inputBuf.array(),
+                    0,
+                    bytesRead
+                )
+
+                destination.writeFully(outputBuf.array(), 0, compressedSize)
+            }
+        } finally {
+            ctx.close()
+            pool.recycle(inputBuf)
+            pool.recycle(outputBuf)
+        }
+    }
 }
 
 @OptIn(DelicateCoroutinesApi::class)
