+++ ktor-utils/posix/src/io/ktor/util/ContentEncodersNative.kt
@@ -24,5 +24,4 @@ public actual object DeflateEncoder : ContentEncoder, Encoder by Identity {
 
 public actual object ZstdEncoder : ContentEncoder, Encoder by Identity {
     actual override val name: String = "zstd"
-    public actual val compressionLevel: Int = 3
 }
